Fixed bug on issue creation
- Jira custom field "category" has been removed. Updated code to
take this into account.
- Deleted maintenance-related workflows as they were based on the
issue category and the maintenance concept has been removed.
